refactor(Education): switch from ES6 class to POJOs
